Chunk based rendering with corrected pixel display + converter
Rendering blocks with data within the c file. Created generate_map() function which creates a flat world. Fixed rendering so it displays up right (had to flip the y axis) so the 0, 0 coordinate is in the bottom left of the world. Chunks are 128 x 64 and only accessed when in a neighbouring chunk for data.
Created a 24 bit to 16 bit converter script to help turn various rgb colour into the hex format for easy texturing.