Added mouse support, Fixed rendering and breaking blocks + formatting
Mouse support can be toggled with switches (9 and 8, corresponding to port 1 and port 2 of the PS2 input). Status indicated on the LEDs.
Fixed rendering of chunks (outside of the first chunk was not rendering correctly, due to overflowing values.
Fixed breaking and placing blocks to work in any chunk.
Plan to rework collision/movement but currently changed the jumping logic to be more realistic.
Changed formatting to use proper tab whitespacing.